+++ src/main/java/Bwead/Bwead.java
@@ -1,3 +1,5 @@
+package Bwead;
+
 import java.io.IOException;
 
 public class Bwead {
