+++ src/main/java/Bwead/BweadException.java
@@ -1,3 +1,5 @@
+package Bwead;
+
 public class BweadException extends RuntimeException {
     public BweadException(String message) {
         super(message);
