+++ src/main/java/Bwead/History.java
@@ -1,3 +1,5 @@
+package Bwead;
+
 import java.util.ArrayList;
 import java.util.Scanner;
 import java.io.File;
