+++ src/main/java/Bwead/Parser.java
@@ -1,3 +1,5 @@
+package Bwead;
+
 import java.util.Scanner;
 import java.time.LocalDate;
