+++ src/main/java/Bwead/Task.java
@@ -1,3 +1,5 @@
+package Bwead;
+
 public class Task {
 
     public boolean done;
