+++ src/main/java/Bwead/Ui.java
@@ -1,3 +1,5 @@
+package Bwead;
+
 import java.util.Scanner;
 import java.io.IOException;
