--- src/main/java/Storage.java
@@ -1,6 +1,0 @@
-public class Storage {
-
-    public Storage(String filePath) {
-
-    }
-}
